Rename Payment to PaymentDomain and add a new PaymentENtity
